get_ini_sections
